cleaned up smolDynamics + fluxes()
fluxes made to be more robust to various bin contents matrices, in case anything gets mixed up in the bin contents matrix. I believe before the code was dependent on the BC matrix being well ordered, which it should be at this point in time but can potentially get mixed up during the step itself. Also changed code to actually free the sims we delete from this

smolDynamics.c was changed to remove commented out code that was used in initial development then abandoned