Changed timing, non-zero KS draft
Timing system changed (clockDouble) to record total time spent in dynamics and split/merge. This code is (mostly) confirmed to work.

Also added additional KS measurements for non-zero flux values. Right now the only thing this does is measure and record this statistic, it does not check a critical value for anything / use this to determine when to stop sim / stop timing.