Alternate Weight Reintroduction
Added second method or reintroducing weight from evacuated flux sims. Prior to now, weights of all replicas were automatically rescaled based on the weight evacuated. Now, the weight from each evacuating replica gets associated with a new sim that is introduced.

This is done through the following method:
First, smolDynamics.c gets a new function added to it, "buildSingleSim" for creating a single Smoldyn simulation.

Second, weSmoldyn.h adds an additional member to the paramsWE struct, this member is defined in WEParams.txt (and loaded with the getParams function). This is currently a binary bit that specifies whether the sims that are evacuated should be replaced or not.

Lastly, the fluxes function gets changed to account for this. If the flag is there, rather than looping through the bin indices associated with the sims and reorganizing / relabeling, the old sim is simply freed and a new sim is created in its place.